pyproject.toml
[project]
name = "pre-commit-ruff-version-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.3",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
ruff-version-sync = "sync:main"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
    "D",      # pydocstyle
    "CPY",    # flake8-copyright
    "T201",   # print found
]

[tool.mypy]
strict = true
allow_redefinition = true
allow_redefinition_new = true
local_partial_types = true
